spike: cut superseded scripts, move bench tooling out of the top level
spike/ had 9 top-level scripts mixed in with the deliverable, so you could not
tell at a glance what was product and what was a one-off experiment. Each script
was written to answer one question during the investigation and never
consolidated afterwards.

  deleted  spike/measure.py     intake-only A/B/C/D, superseded by measure_all.py
  deleted  spike/live_test.py   intake-only live write test, superseded by the
                                real-MCP-client round trip and live_smoke.py
  moved    check_v2, measure_all, live_smoke, probe_model_tokens, compress,
           verify -> spike/bench/

spike/ is now: server_v2.py (entrypoint), v2/ (29 modules), bench/ (tooling).

Fixed sys.path depth and sibling imports for the extra directory level.
Re-verified after the move: check_v2 29/29; measure_all reproduces the full
A/B/C/BD/D table unchanged; compress losslessness 316/316; server launches.

Local spike branch -- not for push.
