+++ spike/bench/compress.py
