+++ spike/bench/measure_all.py
@@ -21,7 +21,7 @@ import os
 import pkgutil
 import sys
 
-sys.path.insert(0, os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
+sys.path.insert(0, os.path.dirname(os.path.dirname(os.path.dirname(os.path.abspath(__file__)))))
 os.environ.setdefault("PLANE_API_KEY", "x")
 os.environ.setdefault("PLANE_WORKSPACE_SLUG", "x")
 
@@ -29,7 +29,7 @@ from fastmcp import FastMCP  # noqa: E402
 
 import spike.v2 as v2pkg  # noqa: E402
 from plane_mcp.server import get_stdio_mcp  # noqa: E402
-from spike.compress import compress  # noqa: E402
+from spike.bench.compress import compress  # noqa: E402
 
 J = dict(separators=(",", ":"))
